Implementacion de roles & permisos

+++ app/Http/Controllers/BookController.php
@@ -7,6 +7,7 @@ use App\Models\Loan;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Auth;
 class BookController extends Controller
 {
     /**
@@ -17,7 +18,8 @@ class BookController extends Controller
     public function index()
     {
         $bookData = Book::all();
-        $loans =Loan::where('state',1)->with('users','books.Category')->get();
+        //$loans =Loan::where('state',1)->with('users','books.Category')->get();
+        $loans =Loan::with('users','books.Category')->get();
         $books = $bookData;
         $categories = Category::all();
 
@@ -30,8 +32,8 @@ class BookController extends Controller
                     $books[$index]['status'] = "0";      
            }
         }
-        //return view('info',compact('books','categories','loans'));
-        return view('books.index',compact('books','categories','loans'));
+        return view('info',compact('books','categories','loans'));
+        //return view('books.index',compact('books','categories','loans'));
     }
 
     /**
@@ -52,35 +54,37 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
-          //Validar los datos del request
-          $validator = Validator::make($request->all(), [
-            'title'=>'required|max:255',
-            'description'=>'required|max:255',
-            'year'=>'required|numeric',
-            'pages'=>'required|numeric',
-            'isbn'=>'required|unique',
-            'editorial'=>'required',
-            'edition'=>'required|numeric',
-            'autor'=>'required',
-            'cover'=>'mimes:jpeg,jpg,png,gif',
-            'category_id' => 'required',
-        ]);
-        //En caso de no ser válidos, se regresa con una respuesta de error
-        if ($validator->fails()) {
-            return  redirect()->back()->with('error', 'Oops! Something went wrong'); 
-        } 
-        if($book = Book::create($request->all())){
-            if ($request->hasFile('cover')) {
-                $file = $request->file('cover');
-                $fileName = 'book_cover'.$book->id.'.'.$file->getClientOriginalExtension();
-                $path = $request->file('cover')->storeAs('img/books',$fileName);
-            }
-            $book->cover = $fileName;
-            $book->save();
-            return  redirect()->back()->with('success', 'Book created successfully');
+        if(Auth::user()->hasPermissionTo('create books')){
+            //Validar los datos del request
+            $validator = Validator::make($request->all(), [
+              'title'=>'required|max:255',
+              'description'=>'required|max:255',
+              'year'=>'required|numeric',
+              'pages'=>'required|numeric',
+              'isbn'=>'required|unique',
+              'editorial'=>'required',
+              'edition'=>'required|numeric',
+              'autor'=>'required',
+              'cover'=>'mimes:jpeg,jpg,png,gif',
+              'category_id' => 'required',
+          ]);
+          //En caso de no ser válidos, se regresa con una respuesta de error
+          if ($validator->fails()) {
+              return  redirect()->back()->with('error', 'Oops! Something went wrong'); 
+          } 
+          if($book = Book::create($request->all())){
+              if ($request->hasFile('cover')) {
+                  $file = $request->file('cover');
+                  $fileName = 'book_cover'.$book->id.'.'.$file->getClientOriginalExtension();
+                  $path = $request->file('cover')->storeAs('img/books',$fileName);
+              }
+              $book->cover = $fileName;
+              $book->save();
+              return  redirect()->back()->with('success', 'Book created successfully');
+          }
+          return  redirect()->back()->with('error', "Sorry, couldn't create book");
         }
-        return  redirect()->back()->with('error', "Sorry, couldn't create book");
-        
+        return redirect()->back()->with("error","You don't have permissions"); 
     }
 
     /**
@@ -113,38 +117,40 @@ class BookController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Book $book)
-    {
-             //Validar los datos del request
-             $validator = Validator::make($request->all(), [
-                'id' => 'required|numeric',
-                'title'=>'required|max:255',
-                'description'=>'required|max:255',
-                'year'=>'required|numeric',
-                'pages'=>'required|numeric',
-                'isbn'=>'required',
-                'editorial'=>'required',
-                'edition'=>'required|numeric',
-                'autor'=>'required',
-                'cover'=>'mimes:jpeg,jpg,png,gif',
-                'category_id' => 'required',
-            ]);
-            //En caso de no ser válidos, se regresa con una respuesta de error
-            if ($validator->fails()) {
-                return  redirect()->back()->with('error', 'Oops! Something went wrong'); 
-            } 
-            $book = Book::find($request['id']);
-            if($book->Update($request->all())){
-                if ($request->hasFile('cover')) {
-                    $file = $request->file('cover');
-                    $fileName = 'book_cover'.$book->id.'.'.$file->getClientOriginalExtension();
-                    $path = $request->file('cover')->storeAs('img/books',$fileName);
-                    $book->cover = $fileName;
-                    $book->save();
-                }
-                return  redirect()->back()->with('success', 'The book has been updated');
-            }
-
-            return redirect()->back()->with('error', 'Sorry, book not updated, try again'); 
+    { 
+        if(Auth::user()->hasPermissionTo('update books')){
+            //Validar los datos del request
+            $validator = Validator::make($request->all(), [
+               'id' => 'required|numeric',
+               'title'=>'required|max:255',
+               'description'=>'required|max:255',
+               'year'=>'required|numeric',
+               'pages'=>'required|numeric',
+               'isbn'=>'required',
+               'editorial'=>'required',
+               'edition'=>'required|numeric',
+               'autor'=>'required',
+               'cover'=>'mimes:jpeg,jpg,png,gif',
+               'category_id' => 'required',
+           ]);
+           //En caso de no ser válidos, se regresa con una respuesta de error
+           if ($validator->fails()) {
+               return  redirect()->back()->with('error', 'Oops! Something went wrong'); 
+           } 
+           $book = Book::find($request['id']);
+           if($book->Update($request->all())){
+               if ($request->hasFile('cover')) {
+                   $file = $request->file('cover');
+                   $fileName = 'book_cover'.$book->id.'.'.$file->getClientOriginalExtension();
+                   $path = $request->file('cover')->storeAs('img/books',$fileName);
+                   $book->cover = $fileName;
+                   $book->save();
+               }
+               return  redirect()->back()->with('success', 'The book has been updated');
+           }
+           return redirect()->back()->with('error', 'Sorry, book not updated, try again'); 
+         }
+         return redirect()->back()->with("error","You don't have permissions");
     }
 
     /**
@@ -155,17 +161,27 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
-        if($book){
-            if($book->delete()){
+        if(Auth::user()->hasPermissionTo('delete books')){
+            if($book){
+                //Se elimian TODOS los préstamos ligados al libro
+                Loan::where('book_id',$book->id)->delete();
+                //Se elimina el libro 
+                if($book->delete()){
+                    return response()->json([
+                        'message' => 'Book deleted successfully', 
+                        'code' => '200'
+                    ]);
+                }
                 return response()->json([
-                    'message' => 'Book deleted successfully', 
-                    'code' => '200'
-                ]);
+                        'message' => "Sorry, coudn't delete the book", 
+                        'code' => '400'
+                    ]);
             }
-            return response()->json([
-                    'message' => "Sorry, coudn't delete the book", 
-                    'code' => '400'
-                ]);
         }
+        return response()->json([
+            'message' => "You don't have permissions", 
+            'code' => '403'
+        ]);
     }
+    
 }
