+++ app/Http/Controllers/CategoryController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
-
+use Auth;
 class CategoryController extends Controller
 {
     /**
@@ -15,7 +15,11 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        //
+        if(Auth::user()->hasPermissionTo('crud categories')){
+            $categories = Category::all();
+            return view('category.index',compact('categories'));
+        }
+        return redirect()->back()->with("error","You don't have permissions");
     }
 
     /**
@@ -36,18 +40,21 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|max:255',
-            'description' => 'required|max:255',
-        ]);
-        if ($validator->fails()) {
-            return  redirect()->back()->with('error', "Invalid fields, couldn't create category");
+        if(Auth::user()->hasPermissionTo('crud categories')){
+            $validator = Validator::make($request->all(), [
+                'name' => 'required|max:255',
+                'description' => 'required|max:255',
+            ]);
+            if ($validator->fails()) {
+                return  redirect()->back()->with('error', "Invalid fields, couldn't create category");
 
+            }
+            if($category = Category::create($request->all())){
+                return  redirect()->back()->with('success', 'Category created successfully');
+            }
+            return  redirect()->back()->with('error', "Sorry, couldn't create category");
         }
-        if($category = Category::create($request->all())){
-            return  redirect()->back()->with('success', 'Category created successfully');
-        }
-        return  redirect()->back()->with('error', "Sorry, couldn't create category");
+        return redirect()->back()->with("error","You don't have permissions");
     }
 
     /**
@@ -81,19 +88,22 @@ class CategoryController extends Controller
      */
     public function update(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|max:255',
-            'description' => 'required|max:255',
-        ]);
-        if ($validator->fails()) {
-            return  redirect()->back()->with('error', "Invalid fields, couldn't edit category");
+        if(Auth::user()->hasPermissionTo('crud categories')){
+            $validator = Validator::make($request->all(), [
+                'name' => 'required|max:255',
+                'description' => 'required|max:255',
+            ]);
+            if ($validator->fails()) {
+                return  redirect()->back()->with('error', "Invalid fields, couldn't edit category");
 
+            }
+            $category = Category::find($request['id']);
+            if($category->Update($request->all()))
+                return  redirect()->back()->with('success', 'Category created successfully');
+            
+            return  redirect()->back()->with('error', "Sorry, couldn't update category");
         }
-        $category = Category::find($request['id']);
-        if($category->Update($request->all()))
-            return  redirect()->back()->with('success', 'Category created successfully');
-        
-        return  redirect()->back()->with('error', "Sorry, couldn't update category");
+        return redirect()->back()->with("error","You don't have permissions");
     }
 
     /**
@@ -104,17 +114,23 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        if($category){
-            if($category->delete()){
+        if(Auth::user()->hasPermissionTo('crud categories')){
+            if($category){
+                if($category->delete()){
+                    return response()->json([
+                        'message' => 'Category deleted successfully', 
+                        'code' => '200'
+                    ]);
+                }
                 return response()->json([
-                    'message' => 'Category deleted successfully', 
-                    'code' => '200'
-                ]);
+                        'message' => "Sorry, coudn't delete category", 
+                        'code' => '400'
+                    ]);
             }
-            return response()->json([
-                    'message' => "Sorry, coudn't delete category", 
-                    'code' => '400'
-                ]);
         }
+        return response()->json([
+            'message' => "You don't have permissions", 
+            'code' => '403'
+        ]);
     }
 }
